+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Custody Service
+  - Add support for limiting the number of position statements returned
 ### Changed
 ### Fixed
 ### Removed